requirements.txt
openai
gradio
faiss-cpu
PyPDF2
pyyaml
python-dotenv
tiktoken
